% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridOperations.R
\name{addManualSimulationGrid}
\alias{addManualSimulationGrid}
\title{Manually Add Grid Search Simulation Agenda}
\usage{
addManualSimulationGrid(opt_object, sim_agenda)
}
\arguments{
\item{opt_object}{An optimization task object as created by \code{initiateOptimizationTask}.}

\item{sim_agenda}{Grid simulation agenda (table). See \code{drawSimulationGrid}.}
}
\value{
The updated \code{opt_object} with a new \code{grid} element and status. The grid is also saved to \code{GridSearch.csv} in the project directory.
}
\description{
Receives a simulation agenda (table) for running Grid Search. Simulation agenda format needs to follow the format drawn by \code{drawSimulationGrid}.
}
