% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridOperations.R
\name{approximateParetoFront}
\alias{approximateParetoFront}
\title{Approximate Pareto Front from Grid Search Result}
\usage{
approximateParetoFront(
  opt_object,
  write_output = T,
  redraw_front = F,
  grid_search_result_filename = NULL,
  output_file_name = NULL
)
}
\arguments{
\item{opt_object}{An optimization task object as created by \code{initiateOptimizationTask}. Must include previously run optimization history.}

\item{write_output}{Boolean prompt to write the Pareto front approximate as output. Filename is 'ParetoFrontApproximate.csv' by default.}

\item{redraw_front}{Boolean prompt to manually require the function to re-draw Pareto front approximate from the latest grid search result.}
}
\value{
The updated \code{opt_object} with a new \code{paretoGrid} element and status. The grid is also saved to \code{ParetoGrid.csv} in the project directory.
}
\description{
Draws a list of Pareto-optimal points from grid search result.
If a previous Pareto front approximate is available in the project directory, this result will be used unless redraw_front is TRUE.
Function returns error message if neither GridSearchResult.csv nor ParetoFrontApproximate.csv was found.
}
\details{
Note: \code{\link{gridSearchEvaluation}} must be run beforehand to generate the grid from which the Pareto front will be approximated from.
}
