% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridOperations.R
\name{drawSimulationGrid}
\alias{drawSimulationGrid}
\title{Draw Grid Search Simulation Agenda}
\usage{
drawSimulationGrid(
  opt_object,
  round_inputs = c(),
  intersect_points = NULL,
  grid_resolution = 50
)
}
\arguments{
\item{opt_object}{An optimization task object as created by \code{initiateOptimizationTask}. Must include a defined \code{searchSpace}.}

\item{round_inputs}{A logical or integer vector indicating which treatment parameters should be rounded to whole numbers. If logical, \code{TRUE} values indicate rounding.}

\item{intersect_points}{Optional named vector. If provided, generates a sub-grid where one or more parameters are fixed at specified values (e.g. the identified optima). Parameters not mentioned in this vector are allowed to vary.}

\item{grid_resolution}{Integer. Number of points to generate per parameter range. Default is 50.}
}
\value{
The updated \code{opt_object} with a new \code{grid} element and status. The grid is also saved to \code{GridSearch.csv} in the project directory.
}
\description{
Generates a grid of treatment parameter combinations for evaluation using with grid search.
Parameter values are created based on the lower and upper bounds defined in the optimization object's
\code{searchSpace}. The generated grid is saved to file and stored in the object.
}
