% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSampling.R
\name{generatePopulationSample}
\alias{generatePopulationSample}
\title{Generate a Population Sample}
\usage{
generatePopulationSample(
  opt_object,
  n_subjects,
  random_effect_parameters,
  random_effect_covariance = NULL,
  patient_demographics_continuous = NULL,
  patient_demographics_categorical = NULL,
  IIV_from_OMEGA = F,
  demographics_from_range = F,
  force_demographics_inrange = F
)
}
\arguments{
\item{opt_object}{mobTD object (a list object containing project configuration).}

\item{n_subjects}{Integer. The number of individuals to include in the population sample.}

\item{random_effect_parameters}{A named numeric vector or list specifying variability parameters. The names must match the ETA identifiers used in the PKPD model included in the opt_object. If a parameter is provided as a percentage (\%IIV), it will be internally converted to ETA scale for simulation. Additionally, any names containing 'iiv' will be automatically renamed to 'eta' to align with the model syntax.}

\item{random_effect_covariance}{Optional matrix. The covariance matrix of the random effects. If NULL, diagonal structure is assumed.}

\item{patient_demographics_continuous}{Optional named list. Each element defines a continuous demographic. If demographics_from_range = TRUE, each element should be a numeric vector of three values: median, lower bound, and upper bound. If demographics_from_range = FALSE, each element should be a numeric vector of two values: mean and standard deviation.}

\item{patient_demographics_categorical}{Optional named list. Each element defines a categorical demographic with level probabilities.}

\item{IIV_from_OMEGA}{Logical. If TRUE, values will be transformed back to OMEGA before resampling. Default is FALSE.}

\item{demographics_from_range}{Logical. If TRUE, continuous demographics will be sampled from specified ranges instead of empirical distributions. Default is FALSE.}

\item{force_demographics_inrange}{Logical. If TRUE and \code{demographics_from_range} is also TRUE, resampling will enforce that values stay within specified ranges. Default is FALSE.}
}
\value{
The updated \code{opt_object} list with the generated population sample. Population sample is also exported in the working directory.
}
\description{
This function generates a population sample by sampling inter-individual variability (IIV) parameters
and patient demographics. Continuous and categorical demographic values can be sampled from
provided ranges or distributions. The resulting sample is stored in the given optimization object,
written to file, and the project is saved.
}
