% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFileOperations.R
\name{loadProject}
\alias{loadProject}
\title{Load Optimization Task Object}
\usage{
loadProject(work_directory, load_all_contents = T)
}
\arguments{
\item{work_directory}{Path to the project folder containing the saved \code{OptimizationTask.Rdata} file.}

\item{load_all_contents}{Logical. If TRUE, large contents such as grid search table, pareto grid, and population samples are reloaded. Default is TRUE.}
}
\value{
The loaded optimization task object.
}
\description{
Loads an optimization task object from a specified working directory. Optionally reloads
large components such as population samples and simulation results.
}
