% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFileOperations.R
\name{modifySimulationSettings}
\alias{modifySimulationSettings}
\title{Modify Simulation Settings}
\usage{
modifySimulationSettings(
  opt_object,
  simulation_tmax,
  simulation_dt,
  n_subjects,
  save_object = T
)
}
\arguments{
\item{opt_object}{An optimization task object as created by \code{initiateOptimizationTask}.}

\item{simulation_tmax}{Numeric. The total simulation time.}

\item{simulation_dt}{Numeric. The time interval between simulation points.}

\item{n_subjects}{Integer. The number of individuals to simulate.}

\item{save_object}{Logical. If TRUE, the updated object will be saved in the working directory, rewriting the previous version. Default is TRUE.}
}
\value{
The updated \code{opt_object}.
}
\description{
Updates the simulation settings in the optimization task object, including simulation duration,
time step, and number of simulated subjects. Optionally saves the updated object to file.
}
