% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFileOperations.R
\name{replicateProject}
\alias{replicateProject}
\title{Replicate Optimization Task}
\usage{
replicateProject(opt_object, new_work_directory)
}
\arguments{
\item{opt_object}{An optimization task object to replicate.}

\item{new_work_directory}{Path to the new directory where the replicated project will be stored.}
}
\value{
The replicated \code{opt_object} with the new working directory.
}
\description{
Creates a copy of an existing optimization task in a new working directory, including
population samples and simulation grids if available. Useful for testing alternative scenarios.
}
