% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFileOperations.R
\name{saveProject}
\alias{saveProject}
\title{Save Optimization Task Object}
\usage{
saveProject(opt_object)
}
\arguments{
\item{opt_object}{An optimization task object as created by \code{initiateOptimizationTask}.}
}
\value{
\code{NULL}. The object is saved to the \code{WorkDirectory} defined in the task object. Previously saved object will be rewritten.
}
\description{
Saves the optimization task object to disk as an \code{.Rdata} file, excluding large components
such as the population sample and simulation grids.
}
